@@ -1,6 +1,6 @@
 <Project>
     <PropertyGroup>
-        <TargetFramework>net8.0</TargetFramework>
+        <TargetFramework>net9.0</TargetFramework>
         <IsPackable>false</IsPackable>
     </PropertyGroup>
 
@@ -19,8 +19,4 @@
         <!-- SYSLIB1006: Multiple logging methods cannot use the same event ID -->
         <NoWarn>$(NoWarn);SYSLIB1006</NoWarn>
     </PropertyGroup>
-
-    <ItemGroup>
-        <PackageReference Include="JetBrains.Annotations" Version="2023.*" PrivateAssets="all" />
-    </ItemGroup>
 </Project>
