@@ -1,16 +1,19 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
+    <Import Project="$([MSBuild]::GetPathOfFileAbove('CVE.Ignore.props', '$(MSBuildThisFileDirectory)../'))"/>
+    
     <PropertyGroup>
         <OutputType>Exe</OutputType>
-        <TargetFramework>net8.0</TargetFramework>
+        <TargetFramework>net9.0</TargetFramework>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
+        <ManagePackageVersionsCentrally>true</ManagePackageVersionsCentrally>
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="BenchmarkDotNet" Version="0.13.12"/>
-        <PackageReference Include="JetBrains.Profiler.Api" Version="1.4.6" />
-        <PackageReference Update="JetBrains.Annotations" Version="2023.3.0"/>
+        <PackageReference Include="BenchmarkDotNet"/>
+        <PackageReference Include="JetBrains.Profiler.Api"/>
+        <PackageReference Update="JetBrains.Annotations"/>
     </ItemGroup>
 
     <ItemGroup>
@@ -18,5 +21,7 @@
         <ProjectReference Include="..\..\src\NexusMods.MnemonicDB\NexusMods.MnemonicDB.csproj"/>
         <ProjectReference Include="..\..\tests\NexusMods.MnemonicDB.TestModel\NexusMods.MnemonicDB.TestModel.csproj"/>
     </ItemGroup>
+    
+    
 
 </Project>
