@@ -1,4 +1,4 @@
-﻿using NexusMods.MnemonicDB.Abstractions.ElementComparers;
+﻿using JetBrains.Annotations;
 using NexusMods.MnemonicDB.Abstractions.ValueSerializers;
 
 namespace NexusMods.MnemonicDB.Abstractions.Attributes;
@@ -6,6 +6,7 @@ namespace NexusMods.MnemonicDB.Abstractions.Attributes;
 /// <summary>
 /// Used to mark the cardinality of an attribute in the database
 /// </summary>
+[PublicAPI]
 public sealed class CardinalityAttribute(string ns, string name) : ScalarAttribute<Cardinality, byte, UInt8Serializer>(ns, name)
 {
     /// <inheritdoc />
