@@ -1,4 +1,5 @@
-﻿using NexusMods.MnemonicDB.Abstractions.IndexSegments;
+﻿using JetBrains.Annotations;
+using NexusMods.MnemonicDB.Abstractions.IndexSegments;
 using NexusMods.MnemonicDB.Abstractions.Models;
 
 namespace NexusMods.MnemonicDB.Abstractions.Attributes;
@@ -6,6 +7,7 @@ namespace NexusMods.MnemonicDB.Abstractions.Attributes;
 /// <summary>
 /// An attribute that represents a collection of values
 /// </summary>
+[PublicAPI]
 public abstract class CollectionAttribute<TValue, TLowLevel, TSerializer>(string ns, string name)
     : Attribute<TValue, TLowLevel, TSerializer>(ns, name, cardinality: Cardinality.Many) 
     where TSerializer : IValueSerializer<TLowLevel>
