@@ -1,4 +1,5 @@
-﻿using NexusMods.MnemonicDB.Abstractions.Models;
+﻿using JetBrains.Annotations;
+using NexusMods.MnemonicDB.Abstractions.Models;
 using NexusMods.MnemonicDB.Abstractions.ValueSerializers;
 
 namespace NexusMods.MnemonicDB.Abstractions.Attributes;
@@ -6,6 +7,7 @@ namespace NexusMods.MnemonicDB.Abstractions.Attributes;
 /// <summary>
 /// An attribute that references another entity.
 /// </summary>
+[PublicAPI]
 public class ReferenceAttribute(string ns, string name) : ScalarAttribute<EntityId, EntityId, EntityIdSerializer>(ns, name)
 {
     /// <inheritdoc />
@@ -18,5 +20,6 @@ public class ReferenceAttribute(string ns, string name) : ScalarAttribute<Entity
 /// <summary>
 /// A typesafe reference attribute, that references entities of type T.
 /// </summary>
+[PublicAPI]
 public sealed class ReferenceAttribute<T>(string ns, string name) : ReferenceAttribute(ns, name)
 where T : IModelDefinition;
