@@ -1,4 +1,4 @@
-﻿using NexusMods.MnemonicDB.Abstractions.ElementComparers;
+﻿using JetBrains.Annotations;
 using NexusMods.MnemonicDB.Abstractions.Models;
 using NexusMods.MnemonicDB.Abstractions.ValueSerializers;
 
@@ -7,6 +7,7 @@ namespace NexusMods.MnemonicDB.Abstractions.Attributes;
 /// <summary>
 /// Represents a collection of references to other entities.
 /// </summary>
+[PublicAPI]
 public class ReferencesAttribute(string ns, string name) : CollectionAttribute<EntityId, EntityId, EntityIdSerializer>(ns, name)
 {
     /// <inheritdoc />
@@ -19,5 +20,6 @@ public class ReferencesAttribute(string ns, string name) : CollectionAttribute<E
 /// <summary>
 /// A typesafe reference attribute, that references entities of type T.
 /// </summary>
+[PublicAPI]
 public sealed class ReferencesAttribute<T>(string ns, string name) : ReferencesAttribute(ns, name)
 where T : IModelDefinition;
