@@ -1,6 +1,5 @@
 using System;
-using NexusMods.MnemonicDB.Abstractions.ElementComparers;
-using NexusMods.MnemonicDB.Abstractions.Models;
+using JetBrains.Annotations;
 using NexusMods.MnemonicDB.Abstractions.ValueSerializers;
 
 namespace NexusMods.MnemonicDB.Abstractions.Attributes;
@@ -8,7 +7,8 @@ namespace NexusMods.MnemonicDB.Abstractions.Attributes;
 /// <summary>
 /// An attribute that holds a timestamp
 /// </summary>
-public class TimestampAttribute(string ns, string name) : ScalarAttribute<DateTimeOffset, long, Int64Serializer>(ns, name)
+[PublicAPI]
+public sealed class TimestampAttribute(string ns, string name) : ScalarAttribute<DateTimeOffset, long, Int64Serializer>(ns, name)
 {
     /// <inheritdoc />
     protected override long ToLowLevel(DateTimeOffset value) => value.ToUnixTimeMilliseconds();
