@@ -1,4 +1,5 @@
-﻿using NexusMods.MnemonicDB.Abstractions.ElementComparers;
+﻿using JetBrains.Annotations;
+using NexusMods.MnemonicDB.Abstractions.ElementComparers;
 using NexusMods.MnemonicDB.Abstractions.ValueSerializers;
 
 namespace NexusMods.MnemonicDB.Abstractions.Attributes;
@@ -6,6 +7,7 @@ namespace NexusMods.MnemonicDB.Abstractions.Attributes;
 /// <summary>
 /// An attribute that represents a value tag value
 /// </summary>
+[PublicAPI]
 public sealed class ValuesTagAttribute(string ns, string name) : ScalarAttribute<ValueTag, byte, UInt8Serializer>(ns, name)
 {
     /// <inheritdoc />
