@@ -1,6 +1,10 @@
 ﻿namespace NexusMods.MnemonicDB.Abstractions.IndexSegments;
 
-public interface IIndexSegment<TValue>
+/// <summary>
+/// An index segment that reutrns values of a specific type
+/// </summary>
+/// <typeparam name="TValue"></typeparam>
+public interface IIndexSegment<out TValue>
 {
     /// <summary>
     /// Gets the value at the given index
