@@ -1,6 +1,7 @@
 ﻿namespace NexusMods.MnemonicDB.Abstractions.Models;
 
-public interface IModelDefinition
-{
-    
-}
+/// <summary>
+/// Marker interface for model definitions, this triggers the soruce generator to generate the model
+/// for the class that implements this interface
+/// </summary>
+public interface IModelDefinition;
